minimize sentry load (maybe) when disabled
i dont think this actually improves performance though :/
